requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
cryptography==43.0.1
debugpy==1.8.5
Deprecated==1.2.14
frozenlist==1.4.1
idna==3.9
multidict==6.1.0
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.11.1
